# docs/scheduler.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "scheduler.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/scheduler/list": {
      "get": {
        "summary": "returns list of scheduled tasks",
        "operationId": "SchedulerApiService_GetScheduled",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/schedulerGetScheduledResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "SchedulerApiService"
        ]
      }
    },
    "/api/scheduler/remove/{id}": {
      "get": {
        "summary": "removes task from execution list",
        "operationId": "SchedulerApiService_Remove",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/schedulerRemoveResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "task ID",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "SchedulerApiService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "schedulerGetScheduledResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/schedulerTask"
          }
        }
      }
    },
    "schedulerRemoveResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "schedulerTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "schedule": {
          "type": "string"
        },
        "skipNext": {
          "type": "boolean"
        },
        "taskType": {
          "$ref": "#/definitions/schedulerTaskType"
        }
      }
    },
    "schedulerTaskType": {
      "type": "object",
      "properties": {
        "of": {
          "type": "string"
        },
        "rule": {
          "type": "string"
        }
      }
    }
  }
}